requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
fastapi
uvicorn
motor
bcrypt
python-jose[cryptography]
passlib[bcrypt]
fastapi-mail
pydantic-settings
python-multipart
python-dotenv
